Add a small style tweak to work descriptions when they are rendered inside the library.
In the library, we shouldn't show the entire work description - the user
should have to click on the work to get the full description. Furthermore,
this makes all work description boxes the same height in the library, which
means that we don't have to use equalizer or any other hack to make
them the same height.