ANNOTATOR PACKAGE!!! YAYYYY
This packages adds a non-standalone browserify bundle (i.e. one that
exports a "require" function) to the app that it is used with. Currently
the following modules are available:

annotator - annotator core
annotator-plugin-highlighter - the highlighter plugin
annotator-store-meteor - a store plugin which works with meteor collections

Technically you can still use the Annotator.Plugin.fetch() API to get plugin
references, but it is not recommended. The meteor store plugin does not register
itself with annotator, and if any non-core plugins (i.e. ones that annotator core
does not register automatically) are ever made available then we will need to either:

1. Add them as entry points in the bundle
2. Require that people "require" them before using the plugin fetch API.

In general, I do not see the point of annotator's plugin registration system and
I want it gone. However, that is going to have to be a project for a later date.
In the mean time, the best thing to do is just to use the "require" function.
This is simple to do and this package will continue to support it into the future.