Say that there are "Hundreds of" books instead of "Over 400".
There are actually only 232 works, so we can only really say "Over 200".
Plus, does anybody really want to be the one responsible for updating
this if we ever have 300 works? I sure don't. "Hundreds" will most likely
continue to be correct, assuming we don't lose a lot of works for
some reason.