Sort annotations by document position!
This works by mucking around with a cursor for the annotations for
a section so that it will be sorted based on a provided comparator
function. This comparator runs xpath-range on a private copy of the
DOM for that section and uses Node.compareDocumentPosition to compare
the positions of the resulting start text nodes.